Move delete variable mutation to variable hooks module
